+++ app/Exceptions/Handler.php
@@ -2,6 +2,8 @@
 
 namespace App\Exceptions;
 
+use App\Events\InternalError;
+use App\Models\ErrorLog;
 use App\Utilities\Response;
 use Exception;
 use Illuminate\Foundation\Exceptions\Handler as ExceptionHandler;
@@ -75,6 +77,22 @@ class Handler extends ExceptionHandler
                         'trace' => $exception->getTraceAsString()
                     ];
                 } else {
+                    try {
+                        $error_data = [
+                            'message' => $exception->getMessage(),
+                            'file' => $exception->getFile(),
+                            'line' => $exception->getLine(),
+                            'trace' => $exception->getTraceAsString()
+                        ];
+
+                        $error = new ErrorLog($error_data);
+                        $error->save();
+
+                        event(new InternalError($error_data));
+                    } catch (Exception $e) {
+                        // Don't worry for now, we just want to try and log some errors
+                    }
+
                     $response = [
                         'message' => 'Sorry, there has been an error, please try again later'
                     ];