+++ app/Providers/EventServiceProvider.php
@@ -18,7 +18,10 @@ class EventServiceProvider extends ServiceProvider
             'App\Listeners\PruneTokens'
         ],
         'App\Events\RequestError' => [
-            'App\Listeners\CaptureAndSend'
+            'App\Listeners\CaptureAndSendRequestError'
+        ],
+        'App\Events\InternalError' => [
+            'App\Listeners\CaptureAndSendInternalError'
         ]
     ];