+++ config/api/item-transfer/fields.php
@@ -8,6 +8,9 @@ return [
         'title' => 'item-transfer/fields.title-resource_id',
         'description' => 'item-transfer/fields.description-resource_id',
         'type' => 'string',
+        'validation' => [
+            'length' => 10
+        ],
         'required' => true
     ]
 ];