+++ config/api/mail/internal-error.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'from_mail' => env('MAIL_FROM_ADDRESS'),
+    'from_name' => env('MAIL_FROM_NAME'),
+    'to' => env('MAIL_TO_ADDRESS'),
+];